requirements.txt
aioredis==1.3.1
aliyun-python-sdk-core==2.13.19
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.8
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==2.2
django-celery==3.3.1
django-cors-headers==3.4.0
django-simpleui==4.0.2
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jmespath==0.10.0
kombu==3.0.37
msgpack==0.6.2
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
pytz==2020.1
redis==2.10.6
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
zope.interface==5.1.0
